@@ -1,23 +1,21 @@
 {
-  "name": "react-timer",
+  "name": "react-todo",
   "version": "1.0.0",
-  "description": "React Timer",
+  "description": "React Todo App",
   "main": "index.js",
   "scripts": {
-    "test": "karma start",
-    "start": "node server.js",
+    "test": "NODE_ENV=test karma start",
+    "start": "npm run build && node server.js",
     "build": "webpack --config webpack.config.js"
   },
   "author": "Juliana Flor",
   "license": "MIT",
+  "engines": {
+    "node": "8.10.0",
+    "npm": "5.6.0"
+  },
   "dependencies": {
     "axios": "^0.18.0",
-    "express": "^4.16.4",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7",
-    "react-router": "^2.0.0"
-  },
-  "devDependencies": {
     "babel-core": "^6.5.1",
     "babel-loader": "^6.2.2",
     "babel-preset-es2015": "^6.5.0",
@@ -26,8 +24,29 @@
     "css-loader": "^0.23.1",
     "deep-freeze-strict": "^1.1.1",
     "expect": "^1.14.0",
+    "express": "^4.16.4",
+    "firebase": "^3.0.2",
     "foundation-sites": "^6.5.1",
     "jquery": ">=3.0.0",
+    "moment": "^2.12.0",
+    "node": "^13.6.0",
+    "node-env-file": "^0.1.8",
+    "node-sass": "^4.13.1",
+    "node-uuid": "^1.4.8",
+    "npm": "^6.13.6",
+    "react": "^0.14.7",
+    "react-addons-test-utils": "^0.14.7",
+    "react-dom": "^0.14.7",
+    "react-redux": "^4.4.1",
+    "react-router": "^2.0.0",
+    "redux": "^3.3.1",
+    "redux-mock-store": "^1.0.3",
+    "redux-thunk": "^2.1.0",
+    "sass-loader": "^7.1.0",
+    "script-loader": "^0.6.1",
+    "style-loader": "^0.13.0"
+  },
+  "devDependencies": {
     "karma": "^0.13.22",
     "karma-chrome-launcher": "^0.2.2",
     "karma-mocha": "^0.2.2",
@@ -35,15 +54,6 @@
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^1.7.0",
     "mocha": "^2.4.5",
-    "moment": "^2.12.0",
-    "node-sass": "^4.11.0",
-    "node-uuid": "^1.4.8",
-    "react-addons-test-utils": "^0.14.7",
-    "react-redux": "^4.4.1",
-    "redux": "^3.3.1",
-    "sass-loader": "^7.1.0",
-    "script-loader": "^0.6.1",
-    "style-loader": "^0.13.0",
-    "webpack": "^1.12.13"
+    "webpack": "^1.15.0"
   }
 }
