@@ -1,14 +1,18 @@
-const redux = require('redux');
-import {searchTextReducer, showCompletedReducer, todosReducer} from 'reducers';
+import * as redux from 'redux';
+import thunk from 'redux-thunk';
 
-export const configure = (initialState = {}) => {
-  const reducer = redux.combineReducers({
+import {searchTextReducer, showCompletedReducer, todosReducer, authReducer} from 'reducers';
+
+export var configure = (initialState = {}) => {
+  var reducer = redux.combineReducers({
     searchText: searchTextReducer,
     showCompleted: showCompletedReducer,
-    todos: todosReducer
+    todos: todosReducer,
+    auth: authReducer
   });
 
-  const store = redux.createStore(reducer, initialState, redux.compose(
+  var store = redux.createStore(reducer, initialState, redux.compose(
+    redux.applyMiddleware(thunk),
     window.devToolsExtension ? window.devToolsExtension() : f => f
   ));
 
