@@ -1,13 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import {Provider} from 'react-redux';
 import TestUtils from 'react-addons-test-utils';
 import $ from 'jquery';
 import expect from 'expect';
-import {Provider} from 'react-redux';
 
-const configureStore = require('configureStore');
-import {TodoList} from 'TodoList';
-import TodoApp from 'TodoApp';
+import * as configureStore from 'configureStore';
+import TodoList from 'TodoList';
+
+import {TodoApp} from 'TodoApp';
 
 describe('TodoApp', () => {
   it('should exist', () => {
@@ -27,4 +28,50 @@ describe('TodoApp', () => {
 
     expect(todoList.length).toEqual(1);
   });
+
+  // it('should add todo to the todos state on handleAddTodo', () => {
+  //   var todoText = 'test text';
+  //   var todoApp = TestUtils.renderIntoDocument(<TodoApp/>);
+
+  //   todoApp.setState({todos: []});
+  //   todoApp.handleAddTodo(todoText);
+
+  //   expect(todoApp.state.todos[0].text).toBe(todoText);
+  //   expect(todoApp.state.todos[0].createdAt).toBeA('number');
+  // });
+
+  // it('should toggle completed value when handleToggle called', () => {
+  //   var todoData = {
+  //     id: 11,
+  //     text: 'Test features',
+  //     completed: false,
+  //     createdAt: 0,
+  //     completedAt: undefined
+  //   };
+  //   var todoApp = TestUtils.renderIntoDocument(<TodoApp/>);
+  //   todoApp.setState({todos: [todoData]});
+
+  //   expect(todoApp.state.todos[0].completed).toBe(false);
+  //   todoApp.handleToggle(11);
+  //   expect(todoApp.state.todos[0].completed).toBe(true);
+  //   expect(todoApp.state.todos[0].completedAt).toBeA('number');
+  // });
+
+  // Test that when toggle from true to false, completedAt get removed
+  // it('should toggle todo from completed to incompoleted', () => {
+  //   var todoData = {
+  //     id: 11,
+  //     text: 'Test features',
+  //     completed: true,
+  //     createdAt: 0,
+  //     completedAt: 123
+  //   };
+  //   var todoApp = TestUtils.renderIntoDocument(<TodoApp/>);
+  //   todoApp.setState({todos: [todoData]});
+
+  //   expect(todoApp.state.todos[0].completed).toBe(true);
+  //   todoApp.handleToggle(11);
+  //   expect(todoApp.state.todos[0].completed).toBe(false);
+  //   expect(todoApp.state.todos[0].completedAt).toNotExist();
+  // });
 });
