@@ -11,56 +11,6 @@ describe('TodoAPI', () => {
     expect(TodoAPI).toExist();
   });
 
-  describe('setTodos', () => {
-    it('should set valid todos array', () => {
-      const todos = [
-        {
-          id: 44,
-          text: 'The text',
-          completed: false
-        }
-      ];
-
-      TodoAPI.setTodos(todos);
-
-      let actualTodos = JSON.parse(localStorage.getItem('todos'));
-
-      expect(actualTodos).toEqual(todos);
-    });
-
-    it('should not set invalid todos array', () => {
-      let badTodos = {
-        a: 'b'
-      };
-
-      TodoAPI.setTodos(badTodos);
-
-      expect(localStorage.getItem('todos')).toBe(null);
-    });
-  });
-
-  describe('getTodos', () => {
-    it('should return empty array for bad localStorage data', () => {
-      let actualTodos = TodoAPI.getTodos();
-
-      expect(actualTodos).toEqual([]);
-    });
-
-    it('should return todos if valid array in localStorage', () => {
-      const todos = [{
-        id: 44,
-        text: 'The text',
-        completed: false
-      }];
-
-      localStorage.setItem('todos', JSON.stringify(todos));
-
-      let actualTodos = TodoAPI.getTodos();
-
-      expect(actualTodos).toEqual(todos);
-    });
-  });
-
   describe('filterTodos', () => {
     const todos = [{
       id: 1,
@@ -92,6 +42,16 @@ describe('TodoAPI', () => {
       let filteredTodos = TodoAPI.filterTodos(todos, true, '');
 
       expect(filteredTodos[0].completed).toBe(false);
+    });
+
+    it('should filter todos by searchText', () => {
+      var filteredTodos = TodoAPI.filterTodos(todos, true, 'text');
+      expect(filteredTodos.length).toBe(2);
+    });
+
+    it('should return all todos if searchText is empty', () => {
+      var filteredTodos = TodoAPI.filterTodos(todos, true, '');
+      expect(filteredTodos.length).toBe(3);
     });
   });
 });
